Move helper functions to kit
