+++ WrightTools/kit/_array.py
@@ -24,6 +24,7 @@ __all__ = [
     "svd",
     "unique",
     "valid_index",
+    "mask_reduce",
 ]
 
 
@@ -338,3 +339,36 @@ def valid_index(index, shape):
         else:
             out.append(i)
     return tuple(out[::-1])
+
+
+def mask_reduce(mask, shape=None):
+    """Reduce a boolean mask, removing all false slices in any dimension.
+    
+    Parameters
+    ----------
+    mask : ndarray with bool dtype
+        The mask which is to be reduced
+    shape : (optional)
+        Further reduce to a shape that broadcasts to the original shape
+
+    Returns 
+    -------
+        A boolean mask with no all False slices.
+    """
+    mask = mask.copy()
+    for i in range(len(mask.shape)):
+        a = mask.copy()
+        j = list(range(len(mask.shape)))
+        j.remove(i)
+        j = tuple(j)
+        a = a.max(axis=j, keepdims=True)
+        idx = [slice(None)] * len(mask.shape)
+        a = a.flatten()
+        idx[i] = [k for k in range(len(a)) if a[k]]
+        mask = mask[tuple(idx)]
+
+    if shape is not None:
+        red = tuple([i for i in range(len(shape)) if shape[i] == 1])
+        return mask.max(axis=red, keepdims=True)
+    return mask
+
